refactored login button into separate file
