+++ client/src/components/google/LoginButton.css
@@ -1,0 +1,6 @@
+#googleButton {
+  margin-left: 25px;
+  margin-right: 25px;
+  padding-bottom: 15px;
+  font-size: 35px;
+}
